requirements.txt
# Core Agent Framework
pyautogen>=0.2.0

# LLM Integration
openai>=1.0.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Web Search & Research Tools
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0

# Evaluation Frameworks
trulens-eval>=0.18.0
deepeval>=0.20.0
pydantic>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Logging & Monitoring
loguru>=0.7.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
